move BattleshipError from utilities to its own errors folder
