+++ server/errors/index.js
@@ -1,5 +1,7 @@
+const BattleshipError = require('./battleship');
 const BattleshipValidationError = require('./validation');
 
 module.exports = {
+    BattleshipError,
     BattleshipValidationError,
 };
